move fhs tests to an easy location
